works: on ios & android, for scrollview & flatlist
