+++ src/containers/NFT/index.ts
@@ -1,0 +1,2 @@
+export * from './NFTHoldings';
+export * from './NFTOverview';
