+++ src/containers/Portfolio/components/Banner.tsx
@@ -1,48 +1,22 @@
 import React, { useContext } from 'react';
 import { useTranslation } from 'react-i18next';
 import { HdWalletContext, PortfolioContext } from '../../../core/util';
-import { ChainHoldings, WalletHoldings, getCurrentChainHoldings } from '../../../core/Portfolio';
-import { Chain } from '../../../constants/server';
+import { getCurrentChainHoldings } from '../../../core/Portfolio';
 import { showToast } from '../../utilities/toastUtility';
 import { setHideBalanceStatus } from '../../../core/asyncStorage';
 import { CyDFastImage, CyDImageBackground, CyDText, CyDTouchView, CyDView } from '../../../styles/tailwindStyles';
 import CyDTokenValue from '../../../components/v2/tokenValue';
 import AppImages from '../../../../assets/images/appImages';
 
-interface BalanceBannerProps {
-  isVerifyCoinChecked: boolean
-  getAllChainBalance: (portfolioState: {
-    statePortfolio: {
-      selectedChain: Chain
-      tokenPortfolio: WalletHoldings
-    }
-  }) => number
+interface BannerProps {
+  checkAllBalance: number | string
 }
 
-export const BalanceBanner = ({ isVerifyCoinChecked, getAllChainBalance }: BalanceBannerProps) => {
+export const Banner = ({ checkAllBalance }: BannerProps) => {
   const { t } = useTranslation();
   const portfolioState = useContext<any>(PortfolioContext);
   const hdWallet = useContext<any>(HdWalletContext);
   const { hideBalance } = hdWallet.state;
-
-  const checkAll = (portfolioState: { statePortfolio: { selectedChain: Chain, tokenPortfolio: WalletHoldings } }) => {
-    if (portfolioState.statePortfolio.selectedChain.backendName !== 'ALL') {
-      const currentChainHoldings = getCurrentChainHoldings(
-        portfolioState.statePortfolio.tokenPortfolio,
-        portfolioState.statePortfolio.selectedChain
-      );
-      if (currentChainHoldings) {
-        const { chainTotalBalance, chainStakedBalance, chainUnbondingBalance, chainUnVerifiedBalance } = currentChainHoldings as ChainHoldings; // Type-assertion (currentChainHoldings can only be of type ChainHoldings if selectedChain.backendName !== 'ALL')
-        return isVerifyCoinChecked
-          ? Number(chainTotalBalance) + Number(chainStakedBalance) + Number(chainUnbondingBalance)
-          : Number(chainTotalBalance) + Number(chainUnVerifiedBalance) + Number(chainStakedBalance) + Number(chainUnbondingBalance);
-      } else {
-        return '...';
-      }
-    } else {
-      return getAllChainBalance(portfolioState);
-    }
-  };
 
   const hideBalances = async () => {
     showToast(hideBalance ? t<string>('PRIVACY_MODE_OFF') : t<string>('PRIVACY_MODE_ON'));
@@ -61,7 +35,7 @@ export const BalanceBanner = ({ isVerifyCoinChecked, getAllChainBalance }: Balan
                 <CyDText>{t('TOTAL_BALANCE')}</CyDText>
                 <CyDView className='flex flex-row items-center py-[3px]'>
                   <CyDTokenValue className='text-[32px] font-extrabold text-primaryTextColor'>
-                    {checkAll(portfolioState)}
+                    {checkAllBalance}
                   </CyDTokenValue>
                   <CyDTouchView onPress={() => {
                     void hideBalances();
