+++ src/containers/Portfolio/components/index.ts
@@ -1,5 +1,6 @@
-export * from './BalanceBanner';
+export * from './Banner';
 export * from './HeaderBar';
 export * from './MessageBanner';
+export * from './RefreshTimerBar';
 export * from './TabBar';
 export * from './TabView';
