+++ src/routes/auth.tsx
@@ -71,13 +71,11 @@ import { CyDImage, CyDTouchView } from '../styles/tailwindStyles';
 import TransactionDetails from '../containers/DebitCard/bridgeCard/transactionDetails';
 import ReferralRewards from '../containers/ReferralRewards/referrals';
 import TokenOverviewV2 from '../containers/TokenOverview';
-import NFTOverviewScreen from '../containers/Portfolio/NFT/NFTOverview';
-import NFTHoldingsScreen from '../containers/Portfolio/NFT/NFTHoldings';
+import { NFTOverviewScreen, NFTHoldingsScreen } from '../containers/NFT';
 import { AddressBook } from '../containers/AddressBook/myAddress';
 import { CreateContact } from '../containers/AddressBook/createContact';
 import TransFiScreen from '../containers/ramp/transFi';
-import { useEffect, useState } from 'react';
-import { useNavigationState } from '@react-navigation/native';
+import { useEffect } from 'react';
 import { BackHandler, Keyboard, StyleProp, TextStyle, ToastAndroid } from 'react-native';
 import ActivateCardScreen from '../containers/DebitCard/bridgeCard/activateCard';
 import SetPinScreen from '../containers/DebitCard/bridgeCard/setPin';
